requirements.txt
langchain==0.3.11
python-dotenv
langchain-community==0.3.11
langchain-core>=0.3.26
langchain-groq
langchain-huggingface
langgraph
langchainhub
pypdf
beautifulsoup4
arxiv
pymupdf
langchain-text-splitters
sentence-transformers
faiss-cpu
langchain-chroma
streamlit
sse-starlette
httpx==0.27.2
orjson
pydantic>=2.0
wikipedia
tavily-python
transformers
huggingface-hub